JIRA:RW-405 case vec.size to int for comparison
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@78323 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objmgr/util/indexer.cpp
@@ -741,7 +741,7 @@ void CBioseqIndex::GetSequence (int from, int to, string& buffer)
             if (from < 0) {
                 from = 0;
             }
-            if (to < 0 || to >= vec.size()) {
+            if (to < 0 || to >= (int) vec.size()) {
                 to = vec.size();
             }
             if (vec.CanGetRange(from, to)) {
@@ -836,7 +836,7 @@ void CFeatureIndex::GetSequence (int from, int to, string& buffer)
             if (from < 0) {
                 from = 0;
             }
-            if (to < 0 || to >= vec.size()) {
+            if (to < 0 || to >= (int) vec.size()) {
                 to = vec.size();
             }
             if (vec.CanGetRange(from, to)) {
